Move test for event endpoint to separate class
